+++ src/main.js
@@ -1,6 +1,11 @@
-import * as commandsHelpers from './commands.js';
 import { COMMANDS, invalidInputMessage } from './consts.js';
-import { catchErrorWrapper, getUserName } from './utils.js';
+import { catchErrorWrapper, getUserName, showAllCommands, getCommandsInfo } from './utils.js';
+import { navigateToHome, goUp, moveToDirectory, showAllDataInDirectory } from './navigation.js';
+import { readFile, createFile, renameFile, copyFile, moveFile, deleteFile } from './file.js';
+import { parseOsArgv } from './os.js';
+import { calculateHash } from './hash.js';
+import { compressFile, decompressFile} from './zip.js';
+
 
 const username = getUserName();
 
@@ -8,8 +13,8 @@ const welcomeMessage = `Welcome to the File Manager, ${username}!\n`;
 const endingMessage = `Thank you for using File Manager, ${username}, goodbye!\n`;
 
 console.log(welcomeMessage);
-commandsHelpers.navigateToHome();
-commandsHelpers.showAllCommands();
+navigateToHome();
+showAllCommands();
 
 process.on('SIGINT', () => {
     console.log(endingMessage);
@@ -17,50 +22,50 @@ process.on('SIGINT', () => {
 });
 
 process.stdin.on("data", async (data) => {
-    const commandInfo = commandsHelpers.getCommandsInfo(data);
+    const commandInfo = getCommandsInfo(data);
 
     switch(commandInfo[0]) {
         case COMMANDS.INFO:
-            commandsHelpers.showAllCommands();
+            showAllCommands();
             break;
         case COMMANDS.UP:
-            commandsHelpers.goUp();
+            goUp();
             break;
         case COMMANDS.CHANGE_DIRECTORY:
-            catchErrorWrapper(() => commandsHelpers.moveToDirectory(commandInfo[1]));
+            catchErrorWrapper(() => moveToDirectory(commandInfo[1]));
             break;    
         case COMMANDS.LIST:
-            catchErrorWrapper(async () => await commandsHelpers.showAllDataInDirectory());
+            catchErrorWrapper(async () => await showAllDataInDirectory());
             break;
         case COMMANDS.CAT:
-            catchErrorWrapper(async () => await commandsHelpers.readFile(commandInfo[1]));
+            catchErrorWrapper(async () => await readFile(commandInfo[1]));
             break;
         case COMMANDS.ADD:
-            catchErrorWrapper(async () => await commandsHelpers.createFile(commandInfo[1]));
+            catchErrorWrapper(async () => await createFile(commandInfo[1]));
             break;
         case COMMANDS.RENAME:
-            catchErrorWrapper(async () => await commandsHelpers.renameFile(commandInfo[1], commandInfo[2]));
+            catchErrorWrapper(async () => await renameFile(commandInfo[1], commandInfo[2]));
             break;
         case COMMANDS.COPY:
-            catchErrorWrapper(async () =>  await commandsHelpers.copyFile(commandInfo[1], commandInfo[2]));
+            catchErrorWrapper(async () =>  await copyFile(commandInfo[1], commandInfo[2]));
             break;
         case COMMANDS.MOVE:
-            catchErrorWrapper(async () => await commandsHelpers.moveFile(commandInfo[1], commandInfo[2]));
+            catchErrorWrapper(async () => await moveFile(commandInfo[1], commandInfo[2]));
             break;
         case COMMANDS.DELETE:
-            catchErrorWrapper(async () => await commandsHelpers.deleteFile(commandInfo[1]));
+            catchErrorWrapper(async () => await deleteFile(commandInfo[1]));
             break;
         case COMMANDS.OS:
-            catchErrorWrapper(() => commandsHelpers.parseOsArgv(commandInfo[1]));
+            catchErrorWrapper(() => parseOsArgv(commandInfo[1]));
             break;
         case COMMANDS.HASH:
-            catchErrorWrapper(() => commandsHelpers.calculateHash(commandInfo[1]));
+            catchErrorWrapper(() => calculateHash(commandInfo[1]));
             break;
         case COMMANDS.COMPRESS:
-            catchErrorWrapper(() => commandsHelpers.compressFile(commandInfo[1], commandInfo[2]));
+            catchErrorWrapper(() => compressFile(commandInfo[1], commandInfo[2]));
             break;
         case COMMANDS.DECOMPRESS:
-            catchErrorWrapper(() => commandsHelpers.decompressFile(commandInfo[1], commandInfo[2]));
+            catchErrorWrapper(() => decompressFile(commandInfo[1], commandInfo[2]));
             break;
         default:
             console.log(invalidInputMessage);
